+++ Gemfile
@@ -1,5 +1,3 @@
 source 'https://rubygems.org'
 
 gemspec
-
-gem "ruby_tree_sitter", "~> 1.9"
