+++ emerge_cli.gemspec
@@ -38,6 +38,7 @@ Gem::Specification.new do |spec|
   spec.add_dependency 'tty-table', '~> 0.12.0'
   spec.add_dependency 'xcodeproj', '~> 1.27.0'
   spec.add_dependency 'ruby_tree_sitter', '~> 1.9'
+  spec.add_dependency 'pry-byebug', '~> 3.8'
 
   spec.add_development_dependency 'minitest', '~> 5.25.1'
   spec.add_development_dependency 'minitest-reporters', '~> 1.7.1'
